Resolving crash errors so that pages load now. Checking functionality
